Fix dashboard loading error
- Fix issue where logged in non-admin users would experience a
flickering dashboard due to a authentication error

#2238
