+++ api/admin/service.js
@@ -292,29 +292,6 @@ async function updateProfile (user, done) {
   });
 }
 
-async function getAnnouncement () {
-  return (await dao.Announcement.find())[0] || {};
-}
-
-async function updateAnnouncement (announcement, userId, done) {
-  announcement.userId = userId;
-  announcement.updatedAt = new Date();
-  if(!announcement.id) {
-    announcement.createdAt = new Date();
-    await dao.Announcement.insert(announcement).then(() => {
-      done(null, true);
-    }).catch (err => {
-      done(err);
-    });
-  } else {
-    await dao.Announcement.update(announcement).then(() => {
-      done(null, true);
-    }).catch (err => {
-      done(err);
-    });
-  }
-}
-
 async function getAgency (id) {
   return await dao.TagEntity.findOne('id = ?', id);
 }
@@ -386,6 +363,4 @@ module.exports = {
   getDashboardTaskMetrics: getDashboardTaskMetrics,
   getActivities: getActivities,
   canAdministerAccount: canAdministerAccount,
-  getAnnouncement: getAnnouncement,
-  updateAnnouncement: updateAnnouncement,
 };
